+++ package/src/back/index.ts
@@ -3,7 +3,9 @@ import { Vector2 } from "../shared/interfaces/Vector2";
 import { MemViewMapper } from "../shared/interfaces/MemViewMapper";
 import { MemViewMapperOutput } from "../shared/interfaces/MemViewMapperOutput";
 import { MemViewElement } from "../shared/interfaces/MemViewElement";
+import { KeyEvent } from "../shared/interfaces/KeyEvent";
 import { Anchor } from "../shared/enums/Anchor";
+import { KeyCode } from "../shared/enums/KeyCode";
 
 export {
   MemView,
@@ -11,5 +13,7 @@ export {
   type MemViewElement,
   type MemViewMapper,
   type MemViewMapperOutput,
+  type KeyEvent,
   Anchor,
+  KeyCode,
 };