+++ package/src/shared/interfaces/KeyEvent.ts
@@ -1,0 +1,6 @@
+import { KeyCode } from "../enums/KeyCode";
+
+export interface KeyEvent {
+  key: KeyCode;
+  isPressed: boolean;
+}